+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Features:
+  * Added support for system Locale type in Java/Swift/Dart.
+
 ## 7.0.5
 Release date: 2020-06-04
 ### Bug fixes: